+++ src/test/java/org/scion/jpan/api/ScmpSenderTest.java
@@ -299,7 +299,7 @@ public class ScmpSenderTest {
 
   private Path getPathToLocalAS_BR() {
     // Border router address
-    return getPathToLocalAS(MockNetwork.BORDER_ROUTER_IPV4, MockNetwork.BORDER_ROUTER_PORT1);
+    return getPathToLocalAS(MockNetwork.BORDER_ROUTER_IPV4, MockNetwork.getBorderRouterPort1());
   }
 
   private Path getPathToLocalAS_BR_30041() {